feat: enhance EarnReward page with new accordion components
- Added ReferFriendsAccordionItem and ReferFriendsListHeader components for improved UI structure.
- Refactored RewardAccountList and RecordsAccordionList to utilize new accordion components for better code organization.
- Adjusted EarnRewardPageWrapper layout for responsive design.
- Updated component exports to include new items for better accessibility.
