+++ packages/kit/src/views/ReferFriends/pages/EarnReward/components/index.ts
@@ -8,3 +8,5 @@ export { RecordsTabSwitcher } from './RecordsTabSwitcher';
 export { PerpsRecordsTab } from './PerpsRecordsTab';
 export { RewardTypeTabs } from './RewardTypeTabs';
 export { UndistributedRewardCard } from './UndistributedRewardCard';
+export { ReferFriendsAccordionItem } from './ReferFriendsAccordionItem';
+export { ReferFriendsListHeader } from './ReferFriendsListHeader';
